Use Gdiffsplit! with fugitive 3.0
Fugitive made a change in how diff splits are rendered in v3.0, where
the default is now a two-pane view. To retain the three-pane view for
our plugin, we need to use the new `:Gdiffsplit!` command instead.

This change leaves a shim where if the 3.0 command isn't defined, we'll
retain the old behavior.